+++ academic_mall.py
@@ -1,0 +1,6 @@
+def enter_academic_mall():
+    print("This is the Academic Mall. " +
+        "To the left, you will see EBB. " + 
+        "To the right, you will see the old part of KHIC " +
+        "and straight ahead you see the new part of KHIC.")
+
